clean up langchain tests
